CIV-12693 - Refactored integration test
