+++ src/integrationTest/java/uk/gov/hmcts/reform/civil/controllers/dashboard/CreateDashboardScenarioTest.java
@@ -18,7 +18,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
 @Testcontainers
-//@Sql("/scripts/dashboardNotifications/create_dashboard_scenarios.sql")
+@Sql("/scripts/dashboardNotifications/create_dashboard_scenarios.sql")
 public class CreateDashboardScenarioTest extends BaseIntegrationTest {
 
     public static final String SCENARIO_HEARING_FEE_PAYMENT_REQUIRED = "scenario.hearing.fee.payment.required";
@@ -67,34 +67,4 @@ public class CreateDashboardScenarioTest extends BaseIntegrationTest {
                 jsonPath("$[0].descriptionEn").value("Pay the hearing fee. <a href=#>Click here</a>")
             );
     }
-
-    @Test
-    void should_create_scenario_for_claim_issue() throws Exception {
-
-        UUID caseId = UUID.randomUUID();
-        String hearingFeeByTime = "4 pm";
-        OffsetDateTime hearingFeeByDate = OffsetDateTime.now();
-        doPost(BEARER_TOKEN,
-               ScenarioRequestParams.builder()
-                   .params(Map.of("defaultRespondTime", hearingFeeByTime,
-                                  "responseDeadline", hearingFeeByDate,
-                                  "daysLeftToRespond", 28,
-                                  "ccdCaseReference", caseId
-                   ))
-                   .build(),
-               DASHBOARD_CREATE_SCENARIO_URL, "Scenario.AAA7.ClaimIssue.Response.Required", caseId
-        )
-            .andExpect(status().isOk());
-
-        //Verify Notification is created
-        doGet(BEARER_TOKEN, GET_NOTIFICATIONS_URL, caseId, "DEFENDANT")
-            .andExpect(status().isOk())
-            .andExpectAll(
-                status().is(HttpStatus.OK.value()),
-                jsonPath("$[0].titleEn").value("You haven´t responded to the claim"),
-                jsonPath("$[0].descriptionEn")
-                    .value("You need to respond before 4 pm on " + hearingFeeByDate + ". There are 28 days remaining. <a href=\"/case/" + caseId + "/response/task-list\">Respond to the claim.</a>.")
-            );
-    }
-
 }
