+++ src/integrationTest/java/uk/gov/hmcts/reform/civil/controllers/dashboard/scenarios/defendant/ClaimIssuedScenarioTest.java
@@ -1,12 +1,15 @@
-package uk.gov.hmcts.reform.civil.controllers.dashboard.scenarios;
+package uk.gov.hmcts.reform.civil.controllers.dashboard.scenarios.defendant;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
+import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.HttpStatus;
 import org.testcontainers.junit.jupiter.Testcontainers;
 import uk.gov.hmcts.reform.civil.controllers.BaseIntegrationTest;
 import uk.gov.hmcts.reform.dashboard.data.ScenarioRequestParams;
 
+import java.time.LocalDate;
+import java.time.OffsetDateTime;
 import java.util.Map;
 import java.util.UUID;
 
@@ -15,67 +18,73 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
 @Testcontainers
-public class DraftClaimScenarioTest extends BaseIntegrationTest {
+public class ClaimIssuedScenarioTest extends BaseIntegrationTest {
 
-    public static final String SCENARIO_DRAFT_CLAIM = "Scenario.AAA7.ClaimIssue.ClaimSubmit.Required";
     private static final String DASHBOARD_CREATE_SCENARIO_URL
         = "/dashboard/scenarios/{scenario_ref}/{unique_case_identifier}";
     private static final String GET_NOTIFICATIONS_URL
         = "/dashboard/notifications/{ccd-case-identifier}/role/{role-type}";
     private static final String GET_TASKS_ITEMS_URL = "/dashboard/taskList/{ccd-case-identifier}/role/{role-type}";
 
+    @MockBean
+    private OffsetDateTime time;
+
     @Test
-    void should_create_draft_claim_scenario() throws Exception {
+    void should_create_scenario_for_claim_issue() throws Exception {
 
         UUID caseId = UUID.randomUUID();
+        String hearingFeeByTime = "4 pm";
+        LocalDate hearingFeeByDate = OffsetDateTime.now().toLocalDate();
         doPost(BEARER_TOKEN,
                ScenarioRequestParams.builder()
-                   .params(Map.of())
+                   .params(Map.of("defaultRespondTime", hearingFeeByTime,
+                                  "responseDeadline", hearingFeeByDate,
+                                  "daysLeftToRespond", 28,
+                                  "ccdCaseReference", caseId
+                   ))
                    .build(),
-               DASHBOARD_CREATE_SCENARIO_URL, SCENARIO_DRAFT_CLAIM, caseId
+               DASHBOARD_CREATE_SCENARIO_URL, "Scenario.AAA7.ClaimIssue.Response.Required", caseId
         )
             .andExpect(status().isOk());
 
         //Verify task Item is created
-        doGet(BEARER_TOKEN, GET_TASKS_ITEMS_URL, caseId, "CLAIMANT")
+        doGet(BEARER_TOKEN, GET_TASKS_ITEMS_URL, caseId, "DEFENDANT")
             .andExpectAll(
                 status().is(HttpStatus.OK.value()),
                 jsonPath("$[0].reference").value(caseId.toString()),
-                jsonPath("$[0].taskNameEn").value("<a>View the claim</a>"),
-                jsonPath("$[0].currentStatusEn").value("Not available yet"),
-                jsonPath("$[1].taskNameEn").value("<a>View information about the claimant</a>"),
-                jsonPath("$[1].currentStatusEn").value("Not available yet"),
+                jsonPath("$[0].taskNameEn").value("<a href={VIEW_CLAIM_URL}>View the claim</a>"),
+                jsonPath("$[0].currentStatusEn").value("Available"),
+                jsonPath("$[1].taskNameEn").value("<a href={VIEW_INFO_ABOUT_CLAIMANT}>View information about the claimant</a>"),
+                jsonPath("$[1].currentStatusEn").value("Available"),
                 jsonPath("$[2].taskNameEn").value("<a>View the response to the claim</a>"),
                 jsonPath("$[2].currentStatusEn").value("Not available yet"),
-                jsonPath("$[3].taskNameEn").value("<a>View information about the defendant</a>"),
-                jsonPath("$[3].currentStatusEn").value("Not available yet"),
+                jsonPath("$[3].taskNameEn").value("<a href={VIEW_INFO_ABOUT_DEFENDANT}>View information about the defendant</a>"),
+                jsonPath("$[3].currentStatusEn").value("Available"),
                 jsonPath("$[4].taskNameEn").value("<a>View hearings</a>"),
                 jsonPath("$[4].currentStatusEn").value("Not available yet"),
                 jsonPath("$[5].taskNameEn").value("<a>Upload hearing documents</a>"),
                 jsonPath("$[5].currentStatusEn").value("Not available yet"),
                 jsonPath("$[6].taskNameEn").value("<a>Add the trial arrangements</a>"),
                 jsonPath("$[6].currentStatusEn").value("Not available yet"),
-                jsonPath("$[7].taskNameEn").value("<a>Pay the hearing fee</a>"),
+                jsonPath("$[7].taskNameEn").value("<a>View the bundle</a>"),
                 jsonPath("$[7].currentStatusEn").value("Not available yet"),
-                jsonPath("$[8].taskNameEn").value("<a>View the bundle</a>"),
-                jsonPath("$[8].currentStatusEn").value("Not available yet"),
-                jsonPath("$[9].taskNameEn").value("<a>View orders and notices</a>"),
+                jsonPath("$[8].taskNameEn").value("<a href={VIEW_ORDERS_AND_NOTICES}>View orders and notices</a>"),
+                jsonPath("$[8].currentStatusEn").value("Available"),
+                jsonPath("$[9].taskNameEn").value("<a>View the judgment</a>"),
                 jsonPath("$[9].currentStatusEn").value("Not available yet"),
-                jsonPath("$[10].taskNameEn").value("<a>View the judgment</a>"),
-                jsonPath("$[10].currentStatusEn").value("Not available yet"),
-                jsonPath("$[11].taskNameEn").value("<a>View applications</a>"),
-                jsonPath("$[11].currentStatusEn").value("Not available yet")
+                jsonPath("$[10].taskNameEn").value("<a>View applications</a>"),
+                jsonPath("$[10].currentStatusEn").value("Not available yet")
 
             );
 
         //Verify Notification is created
-        doGet(BEARER_TOKEN, GET_NOTIFICATIONS_URL, caseId, "CLAIMANT")
+        doGet(BEARER_TOKEN, GET_NOTIFICATIONS_URL, caseId, "DEFENDANT")
             .andExpect(status().isOk())
             .andExpectAll(
                 status().is(HttpStatus.OK.value()),
-                jsonPath("$[0].titleEn").value("This claim has not been submitted"),
-                jsonPath("$[0].descriptionEn").value("Your claim is saved as a draft. <a href=\"{DRAFT_CLAIM_TASK_LIST}\" rel=\"noopener noreferrer\" class=\"govuk-link\">Continue with claim</a>.")
+                jsonPath("$[0].titleEn").value("You haven´t responded to the claim"),
+                jsonPath("$[0].descriptionEn")
+                    .value("You need to respond before 4 pm on " + hearingFeeByDate + ". There are {daysLeftToRespond} days remaining. <a href=\"{RESPONSE_TASK_LIST_URL}\">Respond to the claim.</a>.")
             );
     }
-
 }
